WIP: refactor balances -> custom hook
